poc/requirements.txt
requests
splunk-sdk
python-nmap
colorama
pandas
paramiko
pymongo